feat: added steps for dns setup frontend
